An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done

An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done

An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done

Public fields

df_mixtures

Get the degree of freedom of the block proportion

Active bindings

adjacency

Get the adjacency matrix

mask

Get the mask matrix for dealing with NA

nb_nodes

Get the number of nodes of the level

nb_clusters

Get the number of blocks

distribution

Get the distribution used for the connections

directed

Get if the level is directed or not

mixture_parameter

Access the block proportions

connectivity_parameter

Access the connectivity matrix

membership

Access the variational parameters

entropy

Get the entropy of the model

bound

Get the variational bound of the model

df_mixtures

Get the degree of freedom of the block proportion

df_connect

Get the degree of freedom of the connection parameters

ICL

Get the ICL model selection criterion

penalty

Get the penalty used for computing the ICL

Z

Access the vector of block membership (clustering)

X_hat

Get the connection probability matrix

Methods

Public methods

FitSBM$new()

FitSBM$update_alpha()

FitSBM$update_pi()

FitSBM$init_clustering()

FitSBM$m_step()

FitSBM$ve_step()

FitSBM$do_vem()

FitSBM$permute_empty_class()

FitSBM$clear()

FitSBM$clone()

Method new()

Usage

FitSBM$new(
  Q = 1,
  X = NULL,
  M = NULL,
  directed = FALSE,
  distribution = "bernoulli"
)

Method update_alpha()

Usage

FitSBM$update_alpha(safeguard = 1e-06)

Method update_pi()

Usage

FitSBM$update_pi(safeguard = 1e-06)

Method init_clustering()

Usage

FitSBM$init_clustering(safeguard = 1e-06, method = "hierarchical", Z = NULL)

Method m_step()

Usage

FitSBM$m_step(safeguard = 1e-06)

Method ve_step()

Usage

FitSBM$ve_step(threshold = 1e-06, fixPointIter = 100, safeguard = 1e-06)

Method do_vem()

Usage

FitSBM$do_vem(
  init = "hierarchical",
  threshold = 1e-06,
  maxIter = 1000,
  fixPointIter = 100,
  safeguard = 1e-06,
  Z = NULL,
  bound = NA
)

Method permute_empty_class()

Usage

FitSBM$permute_empty_class()

Method clear()

Usage

FitSBM$clear()

Method clone()

The objects of this class are cloneable with this method.

Usage

FitSBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.