R6Class for multilevel object

Store all simulation parameters and list of fittedmodels. Methods for global inference and model selection are included.

Active bindings

nb_nodes

List of the umber of nodes for each levels

simulation_parameters

List of parameters of the MLVSBM

affiliation_matrix

Access the affiliation matrix

adjacency_matrix

Access the list of adjacency_matrix

memberships

Access the list of the clusterings

fittedmodels

Get the list of selected fitted FitMLVSBM objects

ICL

A summary table of selected fitted models and ICL model selection criterion

ICL_sbm

Summary table of ICL by levels

tmp_fittedmodels

A list of all fitted FitMLVSBM objects

fittedmodels_sbm

A list of selected fitted FitSBM objects of each levels

max_clusters

Access the list of maximum model size

min_clusters

Access the list of minimum model size

directed

Access the list of boolean for levels direction

directed

Access the list of the distribution used for each levels

Methods

Public methods

MLVSBM$estimate_level()

MLVSBM$estimate_sbm_neighbours()

MLVSBM$estimate_sbm_from_neighbours()

MLVSBM$estimate_sbm()

MLVSBM$mcestimate()

MLVSBM$estimate_from_neighbours()

MLVSBM$estimate_neighbours()

MLVSBM$merge_split_membership()

MLVSBM$mc_ms_estimate()

MLVSBM$estimate_one()

MLVSBM$estimate_all_bm()

MLVSBM$new()

MLVSBM$findmodel()

MLVSBM$clearmodels()

MLVSBM$addmodel()

MLVSBM$simulate()

MLVSBM$clone()

Method estimate_level()

Usage

MLVSBM$estimate_level(
  level = "lower",
  Q_min = 1,
  Q_max = 10,
  Z = NULL,
  init = "hierarchical",
  depth = 1,
  nb_cores = NULL
)

Method estimate_sbm_neighbours()

Usage

MLVSBM$estimate_sbm_neighbours(
  level = "lower",
  Q = NULL,
  Q_min = 1,
  Q_max = 10,
  fit = NULL,
  nb_cores = NULL,
  init = NULL
)

Method estimate_sbm_from_neighbours()

Usage

MLVSBM$estimate_sbm_from_neighbours(
  level = "lower",
  Q = NULL,
  fits = NULL,
  nb_cores = NULL
)

Method estimate_sbm()

Usage

MLVSBM$estimate_sbm(level = "lower", Q = Q, Z = NULL, init = "hierarchical")

Method mcestimate()

Usage

MLVSBM$mcestimate(Q, Z = NULL, init = "hierarchical", independent = FALSE)

Method estimate_from_neighbours()

Usage

MLVSBM$estimate_from_neighbours(
  Q,
  models = NULL,
  independent = FALSE,
  nb_cores = nb_cores
)

Method estimate_neighbours()

Usage

MLVSBM$estimate_neighbours(
  Q,
  fit = NULL,
  independent = independent,
  nb_cores = NULL
)

Method merge_split_membership()

Usage

MLVSBM$merge_split_membership(
  fitted = private$fitted[[length(private$fitted)]]
)

Method mc_ms_estimate()

Usage

MLVSBM$mc_ms_estimate(Z = NA, independent = FALSE, nb_cores = NULL)

Method estimate_one()

Usage

MLVSBM$estimate_one(
  Q,
  Z = NULL,
  independent = FALSE,
  init = "hierarchical",
  nb_cores = NULL
)

Method estimate_all_bm()

Usage

MLVSBM$estimate_all_bm(
  Q = NULL,
  Z = NULL,
  independent = FALSE,
  clear = TRUE,
  nb_cores = NULL
)

Method new()

Constructor for R6 class MLVSBM

Usage

MLVSBM$new(
  n = NULL,
  X = NULL,
  A = NULL,
  Z = NULL,
  directed = NULL,
  sim_param = NULL,
  distribution = list("bernoulli", "bernoulli")
)

Arguments

n

A list of size 2, the number of nodes

X

A list of 2 adjacency matrices

A

The affiliation matrix

Z

A list of 2 vectors, the blocks membership

directed

A list of 2 booleans

sim_param

A list of MLVSBM parameters for simulating networks

distribution

The distributions of the interactions ("bernoulli")

Returns

A MLVSBM object

Method findmodel()

Find a fitted model of a given size

Usage

MLVSBM$findmodel(nb_clusters = NA, fit = NA)

Arguments

nb_clusters

A list of the size of the model

fit

if fit = "best" return the best model according to the ICL

Returns

A FitMLVSBM object

Method clearmodels()

delete all fitted models

Usage

MLVSBM$clearmodels()

Method addmodel()

Added a FitMLVSBM object to the list of fitted model

Usage

MLVSBM$addmodel(fit)

Arguments

fit

The FitMLVSBM object to be added

Method simulate()

Usage

MLVSBM$simulate()

Method clone()

The objects of this class are cloneable with this method.

Usage

MLVSBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.