Simulation an adjacency matrix

Simulation an adjacency matrix

simulate_adjacency(
  Z,
  n,
  alpha,
  directed,
  distribution = "bernoulli",
  no_isolated_node = FALSE
)

Arguments

Z
A vector of integer of size n, the label

n
An integer, the number of rows or columbns of the matrix

alpha
A \(\max(Z)\times \max(Z)\) matrix, the connectivity parameters

directed
A boolean, Is the network directed or not ?

distribution
The distribution of the indices: only "bernoulli"

no_isolated_node
A boolean, may row and column of adjacency matrices sum to 0

Value

A nxn adjacency matrix