An R6 Class ocject, a fitted multilevel network once $dovem() is done

An R6 Class ocject, a fitted multilevel network once $dovem() is done

An R6 Class ocject, a fitted multilevel network once $dovem() is done

Active bindings

affiliation_matrix

Get the affiliation matrix

adjacency_matrix

Get the list of adjacency matrices

nb_clusters

Get the list of the number of nodes

nb_clusters

Get the list of the number of blocks

parameters

Get the list of the model parameters

membership

Get the list of the variational parameters

entropy

Get the entropy of the model

bound

Get the variational bound of the model

ICL

Get the ICL model selection criterion of the mdeol

full_penalty

Get the penalty used to compute the ICL

Z

Get the list of block memberships (vector form)

X_hat

Get the list of the matrices of probability conection predictions

map

Get the list of block memberships (matrix form)

Methods

Public methods

FitMLVSBM$new()

FitMLVSBM$update_alpha()

FitMLVSBM$update_pi()

FitMLVSBM$update_gamma()

FitMLVSBM$init_clustering()

FitMLVSBM$clear()

FitMLVSBM$m_step()

FitMLVSBM$ve_step()

FitMLVSBM$do_vem()

FitMLVSBM$permute_empty_class()

FitMLVSBM$clone()

Method new()

Usage

FitMLVSBM$new(
  Q = list(I = 1, O = 1),
  A = NA,
  X = NA,
  M = list(I = NA, O = NA),
  directed = NA,
  distribution = list("bernoulli", "bernoulli"),
  independent = FALSE
)

Method update_alpha()

Usage

FitMLVSBM$update_alpha(safeguard = 2 * .Machine$double.eps)

Method update_pi()

Usage

FitMLVSBM$update_pi(safeguard = 0.01)

Method update_gamma()

Usage

FitMLVSBM$update_gamma(safeguard = 1e-06)

Method init_clustering()

Usage

FitMLVSBM$init_clustering(
  safeguard = 2 * .Machine$double.eps,
  method = "hierarchical",
  Z = NULL
)

Method clear()

Usage

FitMLVSBM$clear()

Method m_step()

Usage

FitMLVSBM$m_step(safeguard = 1e-06)

Method ve_step()

Usage

FitMLVSBM$ve_step(threshold = 1e-06, fixPointIter = 100, safeguard = 1e-06)

Method do_vem()

Usage

FitMLVSBM$do_vem(
  init = "hierarchical",
  threshold = 1e-06,
  maxIter = 1000,
  fixPointIter = 100,
  safeguard = 1e-06,
  Z = NULL
)

Method permute_empty_class()

Usage

FitMLVSBM$permute_empty_class()

Method clone()

The objects of this class are cloneable with this method.

Usage

FitMLVSBM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.